Rename mocks to fixtures
