+++ test/format-commands.test.ts
@@ -6,7 +6,7 @@ import stripAnsi = require('strip-ansi')
 const g: any = global
 g.columns = 80
 import Help from '../src'
-import { AppsDestroy, AppsCreate } from './helpers/mocks'
+import {AppsDestroy, AppsCreate} from './helpers/fixtures'
 
 // extensions to expose method as public for testing
 class TestHelp extends Help {
